requirements.txt
fastapi
uvicorn
litellm
python-dotenv
prometheus-fastapi-instrumentator
requests
streamlit